Move equivocation related logic to EquivocationDetector
